Safari-609.1.11.1/LayoutTests/canvas/philip/tests/2d.line.join.valid.html

<!DOCTYPE html>
<title>Canvas test: 2d.line.join.valid</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

ctx.lineJoin = 'bevel'
_assertSame(ctx.lineJoin, 'bevel', "ctx.lineJoin", "'bevel'");

ctx.lineJoin = 'round';
_assertSame(ctx.lineJoin, 'round', "ctx.lineJoin", "'round'");

ctx.lineJoin = 'miter';
_assertSame(ctx.lineJoin, 'miter', "ctx.lineJoin", "'miter'");


});
</script>
